Expose interop base error code as SqliteResult enum

+++ CiccioSoft.Data.Sqlite/SqliteConnection.cs
@@ -145,7 +145,7 @@ public class SqliteConnection : DbConnection
             }
             catch (SqliteInteropException ex)
             {
-                throw new SqliteException(ex.Message, ex.BaseErrorCode, ex.ExtendedErrorCode, ex);
+                throw new SqliteException(ex.Message, (int)ex.BaseErrorCode, ex.ExtendedErrorCode, ex);
             }
 
             OnStateChange(new StateChangeEventArgs(ConnectionState.Closed, ConnectionState.Open));
